feat: web build target compiles and links via Emscripten
Add the toolchain scaffolding for a WebAssembly + WebGL2 engine build:
- cmake/emscripten-wasm32.cmake toolchain (locates emcc from the devenv
  shell or an emsdk checkout), a 'web' CMake preset (Release, no ImGui),
  scripts/game-build-web.sh, and devenv emscripten + EM_CACHE wiring.
- CMake EMSCRIPTEN guards: glad/backward/enet excluded (GLES3 headers,
  no wasm stack traces, no raw UDP in browsers), Server target skipped,
  physfs shared lib disabled, GAME_WEB define, and wasm link flags
  (WebGL2 only, fixed 512 MB memory, 2 MB stack, FS runtime exports,
  emmalloc backing mimalloc's wasm page source).
- GL portability: gl_headers.h shim (GLES3 on web, glad elsewhere),
  ES 3.0 context on web, a per-platform GLSL prelude ('#version 300 es'
  + precision on web, '#version 410 core' on desktop) prepended at
  compile time, MSAA via multisampled renderbuffer on web (GLES3 has no
  multisample textures; the target is only blit-resolved, never
  sampled), double-precision uniforms uploaded as float on web, and
  guards for GL_MULTISAMPLE/GL_LINE_SMOOTH/KHR_debug.
- Runtime stubs: networking (G.network fails with 'not supported on
  web'), FileWatcher web backend, GetExePath -> /game/, GetUserSaveDir
  -> /save/<app>.
- Portability fixes surfaced by wasm32: Log2 in bits.h computed bit
  width against sizeof(size_t) but counts with a 64-bit clz (broken on
  any 32-bit target), SegmentedList shelf count now respects 32-bit
  size_t, physics_debug_draw and the ImGui includes in debug_ui.h are
  now truly conditional on GAME_WITH_IMGUI, and the no-ImGui DebugUI
  stub mirrors the real FrameContext layout.

Desktop output is unchanged: full test suite passes and both example
games run with byte-equivalent shader sources.

+++ src/bits.h
@@ -15,10 +15,12 @@ inline constexpr size_t NextPow2(size_t n) {
 }
 
 inline constexpr size_t Log2(size_t b) {
+  // The clz builtins operate on 64-bit values regardless of the width of
+  // size_t, so the bit count must match (size_t is 32-bit on wasm).
 #if defined(__GNUC__) || defined(__clang__)
-  return 8 * sizeof(size_t) - __builtin_clzll(b);
+  return 8 * sizeof(unsigned long long) - __builtin_clzll(b);
 #else
-  return 8 * sizeof(size_t) - __lzcnt64(b);
+  return 8 * sizeof(unsigned long long) - __lzcnt64(b);
 #endif
 }
